+++ pkg/schemautils/schemautils.go
@@ -68,6 +68,14 @@ func (schema *ExtendedSchema) GetConciseLocation() string {
 	return location
 }
 
+func (schema *ExtendedSchema) GetHumanReadableLocation() string {
+	location := schema.GetConciseLocation()
+	if location == "" {
+		location = "/"
+	}
+	return location
+}
+
 // Gets the location, excluding ids, including potential parent schemas
 func (schema *ExtendedSchema) GetResolvedLocation() string {
 	location := removeIdFromLocation(schema.Location)
@@ -86,6 +94,17 @@ func (schema *ExtendedSchema) IsProperty() bool {
 	path := strings.Split(location, "/")
 	return len(path) > 1 && path[len(path)-2] == "properties"
 }
+
+func (schema *ExtendedSchema) IsObject() bool {
+	isObject := false
+	for _, t := range schema.Types {
+		if t == "object" {
+			isObject = true
+		}
+	}
+	return isObject
+}
+
 
 func (schema *ExtendedSchema) IsSelfReference() bool {
 	parentLocations := schema.getParentEntryLocations()